video_worker/pyproject.toml
[project]
name = "video-worker"
version = "0.1.0"
description = ""
authors = [
    {name = "Name",email = "username@example.com"}
]

readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "llvmlite (>=0.44.0,<0.45.0)",
    "numba (>=0.61.0,<0.62.0)",
    "openai-whisper (>=20240930,<20240931)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"